Featured datasets CSS improved
