+++ ckanext/opendata_theme/extended_themes/virginiaogda/templates/home/snippets/featured_datasets.html
@@ -4,7 +4,7 @@
   <div class="container">
       <div class="row">
           <div class="col-md-12 span12 col1">
-              <div class="module-content">
+              <div class="module-content featured-dataset">
                 <h2 class="heading"><a href="/dataset?sort=views_recent+desc">{{ h.opendata_theme_get_custom_name('featured-datasets-custom-name', 'Featured Datasets') }}</a></h2>
                 <span class="heading-explanation">Browse featured datasets below.</span>
                 <section>
