+++ ckanext/opendata_theme/extended_themes/virginiaogda/templates/home/snippets/package_list.html
@@ -1,0 +1,11 @@
+{% block package_list %}
+  {% if packages %}
+    <ul class="{{ list_class or 'dataset-list list-unstyled' }}">
+      {% block package_list_inner %}
+        {% for package in packages %}
+          {% snippet 'home/snippets/package_item.html', package=package, item_class=item_class, hide_resources=hide_resources, truncate_title=truncate_title, note_type=note_type %}
+        {% endfor %}
+      {% endblock %}
+    </ul>
+  {% endif %}
+{% endblock %}
