+++ ckanext/opendata_theme/extended_themes/virginiaogda/templates/home/snippets/stats.html
@@ -1,4 +1,5 @@
 {% set stats = h.get_site_statistics() %}
+{% set total_views = h.opendata_theme_get_all_dataset_views() %}
 
 <div class="main-browse">
     <div class="container">
@@ -10,7 +11,7 @@
                     </h2>
                     <div class="cards">
                         <div class="card">
-                                <img class="card-img-top" src="/img/bar-graph.png" alt="Card image cap">
+                                <img class="card-img-top" src="/img/browse.png" alt="Card image cap">
                                 <div class="card-body">
                                     <h5 class="card-title">   {{ h.SI_number_span(stats.dataset_count) }}</h5>
                                     <p class="card-text"> Datasets Available</p>
@@ -18,15 +19,15 @@
                             </div>
 
                             <div class="card">
-                                <img class="card-img-top" src="/browse.png" alt="Card image cap">
+                                <img class="card-img-top" src="/img/bar-graph.png" alt="Card image cap">
                                 <div class="card-body">
-                                    <h5 class="card-title">{{ h.SI_number_span(stats.dataset_count) }}</h5>
+                                    <h5 class="card-title">{{ h.SI_number_span(total_views) }}</h5>
                                     <p class="card-text">Total Datasets Views</p>
                                 </div>
                             </div>
 
                             <div class="card">
-                                <img class="card-img-top" src="/browse.png" alt="Card image cap">
+                                <img class="card-img-top" src="/img/computer-bar-chart.png" alt="Card image cap">
                                 <div class="card-body">
                                     <h5 class="card-title"> {{ h.SI_number_span(stats.dataset_count) }}</h5>
                                     <p class="card-text">Datasets Contributors</p>
